middleware/requirements.txt
fastapi
uvicorn[standard]
httpx[http2]
PyJWT
passlib[bcrypt]
pydantic[email]
aioredis
fastapi-limiter
python-dotenv
loguru
python-multipart
